Limit the version of flint available to 0.5.0 and above

+++ sympy/ntheory/factor_.py
@@ -963,8 +963,7 @@ def _check_termination(factors, n, limitp1, use_trial, use_rho, use_pm1,
         for b, e in facs.items():
             if verbose:
                 print(factor_msg % (b, e))
-            # int() can be removed when https://github.com/flintlib/python-flint/issues/92 is resolved
-            factors[b] = int(exp*e)
+            factors[b] = exp*e
         if verbose:
             print(complete_msg)
         return True
